requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
openapi==1.1.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
pytest-django==4.5.2
pytils==0.4.1
pytz==2022.5
qrcode==7.3.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
